create file qulacs.jl and gate.jl
